Rewrite changeClickedColor function
Rewrote the changeClickedColor function so that instead of relying on
repeated if else statements it uses for of statements instead.

Function also no longer relies on hard coded indexOf search terms,
instead using a value from the childs class list.

It also removed the need to declare a bunch of elements with
querySelector.